+++ electron/preload.js
@@ -7,4 +7,15 @@ contextBridge.exposeInMainWorld('electronAPI', {
   setAutoStart: (enable) => ipcRenderer.invoke('set-auto-start', enable),
   getAutoStart: () => ipcRenderer.invoke('get-auto-start'),
   notify: (title, body) => ipcRenderer.invoke('notify', title, body),
+
+  // Window controls for the custom frameless titlebar
+  windowMinimize: () => ipcRenderer.send('window-minimize'),
+  windowMaximize: () => ipcRenderer.send('window-maximize'),
+  windowClose: () => ipcRenderer.send('window-close'),
+  windowIsMaximized: () => ipcRenderer.invoke('window-is-maximized'),
+  onMaximizeChange: (callback) => {
+    const handler = (_event, isMax) => callback(isMax);
+    ipcRenderer.on('window-maximized-changed', handler);
+    return () => ipcRenderer.removeListener('window-maximized-changed', handler);
+  },
 });